The rain minigame works the way I want.
The rain minigame works the way I want. When clicking in the screen an "A" is generated. By tocuching an "A", this transform into the next letter in the list. If this letter is the next in the list of the secret word, this letter enable rigidbody and collider and moves to an specific place in the screen